fix: restructuring checksum package
