+++ checksum/crc8.go
@@ -8,7 +8,7 @@
 // see crc8_test.go
 
 // Package crc describes algorithms for finding various CRC checksums
-package crc
+package checksum
 
 import "math/bits"
 
@@ -22,8 +22,8 @@ type CRCModel struct {
 	Name   string
 }
 
-// CalculateCRC8 This function calculate CRC8 checksum.
-func CalculateCRC8(data []byte, model CRCModel) uint8 {
+// CRC8 This function calculate CRC8 checksum.
+func CRC8(data []byte, model CRCModel) uint8 {
 	table := getTable(model)
 	crcResult := model.Init
 	crcResult = addBytes(data, model, crcResult, table)
